requirements.txt
Django~=4.0.5
news~=1.0
celery~=5.2.7
django-filter~=22.1
APScheduler~=3.9.1
django-allauth~=0.51.0
django_apscheduler~=0.6.2
redis~=4.3.4
eventlet~=0.33.1

dnspython~=2.2.1
idna~=3.4
requests~=2.28.1
cryptography~=38.0.1
pip~=22.3
wheel~=0.36.2
amqp~=5.1.1
vine~=5.0.0
pytz~=2022.5
kombu~=5.2.4
billiard~=3.6.4.0
Deprecated~=1.2.13
packaging~=21.3
click~=8.1.3
greenlet~=1.1.3.post0
sqlparse~=0.4.3
asgiref~=3.5.2
colorama~=0.4.5
PyJWT~=2.6.0
certifi~=2022.9.24
tzlocal~=4.2
six~=1.16.0
wcwidth~=0.2.5
cffi~=1.15.1
oauthlib~=3.2.2
urllib3~=1.26.12
pyparsing~=3.0.9
pycparser~=2.21
wrapt~=1.14.1
setuptools~=57.0.0
python-dotenv~=0.21.0
django-modeltranslation~=0.18.7